+++ vlmeval/smp/vlm.py
@@ -71,15 +71,6 @@ def decode_base64_to_image_file(base64_string, image_path, target_size=-1):
     image.save(image_path)
 
 
-def LMUDataRoot():
-    if 'LMUData' in os.environ and osp.exists(os.environ['LMUData']):
-        return os.environ['LMUData']
-    home = osp.expanduser('~')
-    root = osp.join(home, 'LMUData')
-    os.makedirs(root, exist_ok=True)
-    return root
-
-
 def build_option_str(option_dict):
     s = 'There are several options: \n'
     for c, content in option_dict.items():
@@ -140,13 +131,3 @@ def circular_pred(df, extract_func=None):
     flag_map = {k: v for k, v in flag_map.items() if valid_map[k]}
     flags = list(flag_map.values())
     return np.mean(flags)
-
-
-def MMBenchOfficialServer():
-    root = LMUDataRoot()
-    for dataset in ['MMBench', 'MMBench_CN', 'MMBench_TEST_EN', 'MMBench_TEST_CN']:
-        if osp.exists(f'{root}/{dataset}.tsv'):
-            data = load(f'{root}/{dataset}.tsv')
-            if 'answer' in data and sum([pd.isna(x) for x in data['answer']]) == 0:
-                return True
-    return False
